paragon id Int to Long change || edit produkt working

+++ app/src/main/java/com/example/photoapp/features/paragon/data/ParagonDao.kt
@@ -18,7 +18,7 @@ interface ParagonDao {
     fun getAll(): List<Paragon>
 
     @Query("SELECT * FROM Paragon WHERE id = :id")
-    fun getById(id: Int): Paragon?
+    fun getById(id: Long): Paragon?
 
     @Insert
     fun insert(paragon: Paragon): Long
